Added all tests for transformation
