+++ config/checkstyle/checkstyle.xml
@@ -29,7 +29,7 @@
 
   <module name="SuppressWarningsFilter" />
   <module name="LineLength">
-    <property name="max" value="160"/>
+    <property name="max" value="130"/>
     <property name="ignorePattern" value="^package.*|^import.*|a href|href|http://|https://|ftp://"/>
   </module>
   <module name="TreeWalker">
