+++ src/functionalTest/resources/responses/bulk-scan-c100-transform-output.json
@@ -112,7 +112,7 @@
       },
       "personWhoLivesWithChild": "living other details",
       "applicationPermissionRequired": "yes",
-      "habitualResidentINotherState": "yes",
+      "habitualResidentInOtherState": "yes",
       "parentalResponsibilityDetails": "parentResponsibility",
       "familyMediatorMiam": null,
       "isCaseUrgent": "yes",
@@ -236,7 +236,20 @@
             "addressLivedLessThan5YearsDetails": "227, Topsfield Parade, N8 8PE"
           }
         }
-      ]
+      ],
+      "internationalElementTable": {
+        "habitualResidentInOtherState": "yes",
+        "habitualResidentInOtherStateGiveReason": null,
+        "jurisdictionIssue": "Yes",
+        "jurisdictionIssueGiveReason": "International jurisdiction big issue",
+        "requestToForeignAuthority": null,
+        "requestToForeignAuthorityGiveReason": null
+      },
+      "litigationCapacityTable": {
+        "litigationCapacityFactors": "Factors affecting litigation capacity",
+        "litigationCapacityReferrals": "Assessment by adult learning team",
+        "litigationCapacityOtherFactors": "Factors affecting person in court"
+      }
     }
   },
   "warnings": []
