+++ src/integrationTest/resources/request/bulk-scan-c100-validate-input.json
@@ -500,6 +500,46 @@
     {
       "name": "respondent2AllAddressesForLastFiveYears",
       "value": "227, Topsfield Parade, N8 8PE"
+    },
+    {
+      "name": "international_or_factorsAffectingLitigation",
+      "value": "yes"
+    },
+    {
+      "name": "internationalElement_Resident_of_another_state",
+      "value": "No"
+    },
+    {
+      "name": "internationalElement_Resident_of_another_state_details",
+      "value": ""
+    },
+    {
+      "name": "internationalElement_jurisdictionIssue",
+      "value": "Yes"
+    },
+    {
+      "name": "withoutNotice_jurisdictionIssue_details",
+      "value": "International jurisdiction big issue"
+    },
+    {
+      "name": "internationalElement_request_toCentral_or_Consular_authority",
+      "value": ""
+    },
+    {
+      "name": "internationalElement_request_toCentral_or_Consular_authority_details",
+      "value": ""
+    },
+    {
+      "name": "factorAffectingLitigationCapacity",
+      "value": "Factors affecting litigation capacity"
+    },
+    {
+      "name": "assessmentByAdultLearningTeam",
+      "value": "Assessment by adult learning team"
+    },
+    {
+      "name": "factorsAffectingPersonInCourt",
+      "value": "Factors affecting person in court"
     }
   ]
 }
